# cmd/user-portal/docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: Code - HTTP код ошибки
        type: integer
      error:
        description: Error - код ошибки
        type: string
      message:
        description: Message - сообщение об ошибке
        type: string
      timestamp:
        description: Timestamp - время возникновения ошибки
        type: string
    type: object
...
  models.ListFilesResponse:
    properties:
      files:
        description: Files - список файлов
        items:
          $ref: '#/definitions/models.UploadedFile'
        type: array
      page:
        description: Page - номер текущей страницы
        type: integer
      pageSize:
        description: PageSize - размер страницы
        type: integer
      total:
        description: Total - общее количество файлов
        type: integer
    type: object
...
  models.RequestPasswordResetRequest:
    properties:
      email:
        description: Email - адрес электронной почты для отправки кода сброса
        example: user@example.com
        type: string
    required:
    - email
    type: object
  models.RequestPasswordResetResponse:
    properties:
      message:
        description: Message - статусное сообщение
        example: Password reset code sent to your email
        type: string
      token_id:
        description: TokenID - идентификатор сгенерированного токена сброса
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  models.ResetPasswordRequest:
    properties:
      code:
        description: Code - 6-значный код подтверждения
        example: "123456"
        type: string
      new_password:
        description: NewPassword - новый пароль (минимум 8 символов)
        example: newpassword123
        minLength: 8
        type: string
      token_id:
        description: TokenID - идентификатор токена сброса
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - code
    - new_password
    - token_id
    type: object
  models.ResetPasswordResponse:
    properties:
      message:
        description: Message - статусное сообщение
        example: Password reset successfully
        type: string
    type: object
...
  models.TranscriptionStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-comments:
      StatusCompleted: Завершено
      StatusFailed: Ошибка
      StatusPending: Ожидает обработки
      StatusProcessing: В процессе обработки
    x-enum-descriptions:
    - Ожидает обработки
    - В процессе обработки
    - Завершено
    - Ошибка
    x-enum-varnames:
    - StatusPending
    - StatusProcessing
    - StatusCompleted
    - StatusFailed
...
  models.UpdateProfileRequest:
    properties:
      avatar:
        description: Avatar - URL аватара
        example: https://example.com/avatar.jpg
        type: string
      bio:
        description: Bio - биография
        example: Software developer
        type: string
      first_name:
        description: FirstName - имя
        example: John
        type: string
      language:
        description: Language - предпочитаемый язык
        example: en
        type: string
      last_name:
        description: LastName - фамилия
        example: Doe
        type: string
      phone:
        description: Phone - номер телефона
        example: "+1234567890"
        type: string
    type: object
  models.UploadAvatarResponse:
    properties:
      avatar_url:
        description: AvatarURL - URL загруженного аватара
        example: https://example.com/avatars/user-123.jpg
        type: string
      message:
        description: Message - статусное сообщение
        example: Avatar uploaded successfully
        type: string
    type: object
...
  models.UploadedFile:
    properties:
      file_size:
        description: FileSize - размер файла в байтах
        type: integer
      filename:
        description: Filename - имя файла в хранилище
        type: string
      group_id:
        description: GroupID - идентификатор группы
        type: string
      id:
        description: ID - уникальный идентификатор файла
        type: string
      metadata:
        additionalProperties: true
        description: Metadata - дополнительные метаданные файла
        type: object
      mime_type:
        description: MimeType - MIME тип файла
        type: string
      original_name:
        description: OriginalName - оригинальное имя файла
        type: string
      processed_at:
        description: ProcessedAt - время завершения обработки файла
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.TranscriptionStatus'
        description: Status - статус обработки файла
      storage_path:
        description: StoragePath - путь к файлу в хранилище
        type: string
      transcription_id:
        description: TranscriptionID - идентификатор транскрипции (если есть)
        type: string
      uploaded_at:
        description: UploadedAt - время загрузки файла
        type: string
      user_id:
        description: UserID - идентификатор пользователя, загрузившего файл
        type: string
    type: object
  models.UserInfo:
    properties:
      avatar:
        description: Avatar - URL аватара или base64 изображение
        example: https://example.com/avatar.jpg
        type: string
      bio:
        description: Bio - биография пользователя
        example: Software developer
        type: string
      department_id:
        description: DepartmentID - идентификатор отдела, к которому принадлежит пользователь
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      email:
        description: Email - адрес электронной почты пользователя
        example: admin@recontext.online
        type: string
      first_name:
        description: FirstName - имя пользователя
        example: John
        type: string
      id:
        description: ID - уникальный идентификатор пользователя
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      language:
        description: Language - предпочитаемый язык пользователя
        example: en
        type: string
      last_name:
        description: LastName - фамилия пользователя
        example: Doe
        type: string
      permissions:
        allOf:
        - $ref: '#/definitions/models.UserPermissions'
        description: Permissions - специфические разрешения пользователя
      phone:
        description: Phone - номер телефона пользователя
        example: "+1234567890"
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        description: Role - роль пользователя в системе
        example: admin
      username:
        description: Username - имя пользователя для входа
        example: admin
        type: string
    type: object
  models.UserPermissions:
    properties:
      can_approve_recordings:
        description: CanApproveRecordings - разрешение на утверждение записей
        type: boolean
      can_manage_department:
        description: CanManageDepartment - разрешение на управление отделом
        type: boolean
      can_schedule_meetings:
        description: CanScheduleMeetings - разрешение на планирование видеовстреч
        type: boolean
    type: object
  models.UserRole:
    enum:
    - admin
    - user
    - operator
    - service
    type: string
    x-enum-comments:
      RoleService: For service-to-service authentication
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - For service-to-service authentication
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
    - RoleOperator
    - RoleService
  models.VerifyResetCodeRequest:
    properties:
      code:
        description: Code - 6-значный код подтверждения
        example: "123456"
        type: string
      token_id:
        description: TokenID - идентификатор токена сброса для проверки
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - code
    - token_id
    type: object
  models.VerifyResetCodeResponse:
    properties:
      message:
        description: Message - дополнительная информация о проверке
        example: Code verified successfully
        type: string
      valid:
        description: Valid - валиден ли код
        example: true
        type: boolean
    type: object
host: localhost:8081
info:
  contact:
    email: support@recontext.online
    name: API Support
    url: http://recontext.online/support
  description: 'API пользовательского портала Recontext.online: загрузка записей,
    транскрибация, поиск и управление материалами'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://recontext.online/terms/
  title: Recontext.online User Portal API
  version: 0.1.0
paths:
...
  /api/v1/auth/password-reset/request:
    post:
      consumes:
      - application/json
      description: Запросить код сброса пароля для отправки по электронной почте
      parameters:
      - description: Адрес электронной почты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RequestPasswordResetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Запросить сброс пароля
      tags:
      - Auth
  /api/v1/auth/password-reset/reset:
    post:
      consumes:
      - application/json
      description: Сбросить пароль пользователя с использованием проверенного кода
      parameters:
      - description: Данные для сброса пароля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Сбросить пароль с кодом
      tags:
      - Auth
  /api/v1/auth/password-reset/verify:
    post:
      consumes:
      - application/json
      description: Проверить, что предоставленный код сброса действителен
      parameters:
      - description: Идентификатор токена и код
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyResetCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VerifyResetCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Проверить код сброса пароля
      tags:
      - Auth
...
  /api/v1/files:
    get:
      description: Получить постраничный список загруженных файлов пользователя
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListFilesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список загруженных файлов
      tags:
      - Files
...
  /api/v1/users:
    get:
      description: Получить список пользователей для выбора участников встречи
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Список пользователей для участников встречи
      tags:
      - Helpers
  /api/v1/users/{id}:
    get:
      description: Получить информацию о профиле пользователя
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить профиль пользователя
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Обновить информацию профиля пользователя (самообслуживание)
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Данные профиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль пользователя
      tags:
      - Profile
  /api/v1/users/{id}/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Загрузить изображение аватара пользователя
      parameters:
      - description: Файл изображения аватара
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UploadAvatarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить аватар пользователя
      tags:
      - Profile
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"